Moved system functionality to a trait.
